project-{topic}-enhanced/requirements.txt
```
fastapi[all]>=0.104.1
uvicorn>=0.23.2
SQLAlchemy>=2.0.22
asyncpg>=0.28.0
psycopg2-binary>=2.9.9 # For Alembic/local dev with sync connections if needed
alembic>=1.12.0
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
redis>=5.0.1
fastapi-cache2[redis]>=0.2.1
fastapi-limiter>=0.1.5
httpx>=0.25.1 # For API testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
locust>=2.17.0
```